agentlego/requirements_all.txt
accelerate==0.29.3
addict==2.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.2
annotated-types==0.6.0
anyio==4.3.0
Brotli 
certifi 
cffi==1.16.0
charset-normalizer 
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
diffusers==0.27.2
distro==1.9.0
easyocr==1.7.1
einops==0.7.0
fastapi==0.110.2
filelock 
fonttools==4.51.0
fsspec==2024.3.1
func-timeout==4.3.5
gmpy2 
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
idna 
imageio==2.34.1
importlib_metadata==7.1.0
Jinja2 
jmespath==0.10.0
joblib==1.4.0
kiwisolver==1.4.5
langid==1.1.6
lazy_loader==0.4
makefun==1.15.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe 
mat4py==0.6.0
matplotlib==3.8.4
mdurl==0.1.2
mkl-fft 
mkl-random 
mkl-service
mmdet==3.3.0
mmengine==0.10.4
mmpretrain==1.2.0
model-index==0.1.11
modelindex==0.0.2
mpmath 
networkx 
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.4
openai==1.23.3
openapi-pydantic==0.4.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.38
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==2.2.2
pillow 
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
PySocks 
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.4
PyYAML 
rapidfuzz==3.8.1
regex==2024.4.16
requests==2.28.2
rich==13.4.2
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
segment-anything==1.0
sentence-transformers==2.7.0
sentencepiece==0.2.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy 
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
thefuzz==0.22.1
threadpoolctl==3.4.0
tifffile==2024.4.18
tokenizers==0.19.1
tomli==2.0.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.65.2
transformers==4.40.1
transformers-stream-generator==0.0.5
triton==2.1.0
typer==0.12.3
typing_extensions 
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
yapf==0.40.2
zipp==3.18.1
